+++ openfisca_core/periods/__init__.py
@@ -22,7 +22,7 @@
 # See: https://www.python.org/dev/peps/pep-0008/#imports
 
 from . import types
-from ._errors import InstantError, ParserError
+from ._errors import InstantError, ParserError, PeriodError
 from .config import (
     INSTANT_PATTERN,
     date_by_instant_cache,
@@ -62,6 +62,7 @@ __all__ = [
     "MONTH",
     "ParserError",
     "Period",
+    "PeriodError",
     "WEEK",
     "WEEKDAY",
     "YEAR",