+++ openfisca_core/periods/_parsers.py
@@ -1,19 +1,64 @@
+"""To parse periods and instants from strings."""
+
 from __future__ import annotations
 
-import re
+import datetime
 
 import pendulum
 
 from . import types as t
-from ._errors import ParserError
+from ._errors import InstantError, ParserError, PeriodError
 from .date_unit import DateUnit
 from .instant_ import Instant
 from .period_ import Period
 
-invalid_week = re.compile(r".*(W[1-9]|W[1-9]-[0-9]|W[0-5][0-9]-0)$")
+
+def parse_instant(value: str) -> t.Instant:
+    """Parse a string into an instant.
+
+    Args:
+        value (str): The string to parse.
+
+    Returns:
+        An InstantStr.
+
+    Raises:
+        InstantError: When the string is not a valid ISO Calendar/Format.
+        ParserError: When the string couldn't be parsed.
+
+    Examples:
+        >>> parse_instant("2022")
+        Instant((2022, 1, 1))
+
+        >>> parse_instant("2022-02")
+        Instant((2022, 2, 1))
+
+        >>> parse_instant("2022-W02-7")
+        Instant((2022, 1, 16))
+
+        >>> parse_instant("2022-W013")
+        Traceback (most recent call last):
+        openfisca_core.periods._errors.InstantError: '2022-W013' is not a va...
+
+        >>> parse_instant("2022-02-29")
+        Traceback (most recent call last):
+        pendulum.parsing.exceptions.ParserError: Unable to parse string [202...
+
+    """
+
+    if not isinstance(value, t.InstantStr):
+        raise InstantError(str(value))
+
+    date = pendulum.parse(value, exact=True)
+
+    if not isinstance(date, datetime.date):
+        msg = f"Unable to parse string [{value}]"
+        raise ParserError(msg)
+
+    return Instant((date.year, date.month, date.day))
 
 
-def parse_period(value: object) -> None | t.Period:
+def parse_period(value: str) -> t.Period:
     """Parses ISO format/calendar periods.
 
     Such as "2012" or "2015-03".
@@ -30,34 +75,13 @@ def parse_period(value: object) -> None | t.Period:
 
     """
 
-    # If it is not a string, next!
-    if not isinstance(value, str):
-        msg = f"Expected a {str.__name__}, got {type(value).__name__}."
-        raise NotImplementedError(msg)
+    try:
+        instant = parse_instant(value)
 
-    # If it's a complex period, next!
-    if len(value.split(":")) != 1:
-        return None
-
-    # Check for a non-empty string.
-    if len(value) == 0:
-        raise AttributeError
-
-    # If it's negative, next!
-    if value[0] == "-":
-        raise ValueError
-
-    # If it's an invalid week, next!
-    if invalid_week.match(value):
-        raise ParserError
+    except InstantError as error:
+        raise PeriodError(value) from error
 
     unit = parse_unit(value)
-    date = pendulum.parse(value, exact=True)
-
-    if not isinstance(date, pendulum.Date):
-        raise TypeError
-
-    instant = Instant((date.year, date.month, date.day))
 
     return Period((unit, instant, 1))
 
@@ -72,37 +96,26 @@ def parse_unit(value: str) -> t.DateUnit:
         A DateUnit.
 
     Raises:
-        ValueError when no DateUnit can be determined.
+        InstantError: when no DateUnit can be determined.
 
     Examples:
         >>> parse_unit("2022")
         <DateUnit.YEAR: 'year'>
 
-        >>> parse_unit("2022-W03-01")
+        >>> parse_unit("2022-W03-1")
         <DateUnit.WEEKDAY: 'weekday'>
 
     """
 
-    format_y, format_ym, format_ymd = 1, 2, 3
+    if not isinstance(value, t.InstantStr):
+        raise InstantError(str(value))
+
     length = len(value.split("-"))
-    isweek = value.find("W") != -1
 
-    if length == format_y:
-        return DateUnit.YEAR
+    if isinstance(value, t.ISOCalendarStr):
+        return DateUnit.isocalendar[-length]
 
-    if length == format_ym:
-        if isweek:
-            return DateUnit.WEEK
-
-        return DateUnit.MONTH
-
-    if length == format_ymd:
-        if isweek:
-            return DateUnit.WEEKDAY
-
-        return DateUnit.DAY
-
-    raise ValueError
+    return DateUnit.isoformat[-length]
 
 
-__all__ = ["parse_period", "parse_unit"]
+__all__ = ["parse_instant", "parse_period", "parse_unit"]